move log module to sqs listener
