+++ lib/sqs/log/event.js
