+++ lib/sqs/serverless/dispatcher.js
@@ -2,7 +2,7 @@
 
 const SQSServerlessHandlerError = require('./error');
 
-const EventLogger = require('../../modules/log/event');
+const EventLogger = require('../../sqs/log/event');
 
 class ServerlessDispatcher {
